feat: Adds basic validation to JobService

+++ src/api-gateway/src/job/job.service.ts
@@ -33,6 +33,7 @@ import { JobTagRepository } from './job-tag.repository';
 import { CreatePriorityTagDto, CreateTagDto } from './dto/create-tag.dto';
 import { JobPriorityTag, JobTag } from './entities/job-tag.entity';
 import { DeleteTagDto } from './dto/edit-tag.dto';
+import { Employee } from '../employee/entities/employee.entity';
 
 @Injectable()
 export class JobService {
@@ -147,60 +148,6 @@ export class JobService {
   async softDelete(id: Types.ObjectId): Promise<boolean> {
     await this.jobRepository.delete(id);
     return true;
-  }
-
-  async jobIsValid(job: Job): Promise<ValidationResult> {
-    if (!job) {
-      return new ValidationResult(false, 'Job is null');
-    }
-
-    if (!job.companyId || !job.assignedBy) {
-      return new ValidationResult(false, 'CompanyId or assignedBy is invalid');
-    }
-
-    if (!job.companyId || !job.assignedBy) {
-      return new ValidationResult(false, 'CompanyId or assignedBy is invalid');
-    }
-
-    const exists = await this.employeeService.employeeExists(job.assignedBy);
-    const isInCompany = await this.companyService.employeeIsInCompany(
-      job.companyId,
-      job.assignedBy,
-    );
-    if (!exists || !isInCompany) {
-      return new ValidationResult(
-        false,
-        'Assigned By is invalid or Employee is not in company',
-      );
-    }
-
-    if (job.assignedEmployees) {
-      for (const employee of job.assignedEmployees.employeeIds) {
-        const exists = await this.employeeService.employeeExists(employee);
-        if (!exists) {
-          return new ValidationResult(false, `Employee: ${employee} Not found`);
-        }
-      }
-    }
-
-    if (job.clientId) {
-      const exists = await this.clientService.clientExists(job.clientId);
-      if (!exists) {
-        return new ValidationResult(false, 'Client does not exist');
-      }
-    }
-
-    if (job.companyId) {
-      const exists = await this.companyService.companyIdExists(job.companyId);
-      if (!exists) {
-        return new ValidationResult(
-          false,
-          `Company: ${job.companyId} Not found`,
-        );
-      }
-    }
-
-    return new ValidationResult(true);
   }
 
   async jobCreateIsValid(job: CreateJobDto): Promise<ValidationResult> {
@@ -375,54 +322,59 @@ export class JobService {
       throw new UnauthorizedException('User not in Same Company');
     }
 
-    return this.jobRepository.findAllForEmployee(employeeId);
+    return this.jobRepository.findAllForEmployeeDetailed(employeeId);
   }
 
-  async getAllForEmployeeDetailed(
-    userId: Types.ObjectId,
-    jobId: Types.ObjectId,
-  ) {
+  async assignEmployee(userId: Types.ObjectId, jobAssignDto: JobAssignDto) {
     ///Validation
-    const user = await this.usersService.getUserById(userId);
-    const job = await this.getJobById(jobId);
+    await this.userIdMatchesEmployeeId(userId, jobAssignDto.employeeId);
 
-    if (user == null) throw new NotFoundException('User not found');
-    if (job == null) throw new NotFoundException('Job not found');
-    const relatedCompany = user.joinedCompanies
-      .filter((x) => x.companyId.toString() === job.companyId.toString())
-      .pop(); //sus, but users must only have one employee per company
-    /*    let valid = false;
-    for (const joinedCompany of user.joinedCompanies) {
-      if (
-        job.assignedBy.toString() == joinedCompany.employeeId.toString() ||
-        job.assignedEmployees.employeeIds.includes(joinedCompany.employeeId)
-      ) {
-        valid = true;
-        break;
-      }
-    }
-    if (!valid) throw new NotAcceptableException('user not related to job');*/
-    ///
+    const job = await this.getJobById(jobAssignDto.jobId);
+    if (!job) throw new NotFoundException('Job not found');
+    const alreadyAssigned = job.assignedEmployees.employeeIds.some(
+      (id) => id.toString() === jobAssignDto.employeeToAssignId.toString(),
+    );
+    if (alreadyAssigned) throw new ConflictException('Already Assigned');
+    /// Role-based stuff
+    //TODO: Implement later
 
-    //
-    return this.jobRepository.findAllForEmployeeDetailed(
-      relatedCompany.employeeId,
+    return await this.jobRepository.assignEmployee(
+      jobAssignDto.employeeToAssignId,
+      jobAssignDto.jobId,
     );
   }
 
-  async assignEmployee(
+  async unassignEmployee(userId: Types.ObjectId, jobAssignDto: JobAssignDto) {
+    ///Validation
+    await this.userIdMatchesEmployeeId(userId, jobAssignDto.employeeId);
+
+    const job = await this.getJobById(jobAssignDto.jobId);
+    if (!job) throw new NotFoundException('Job not found');
+    const alreadyAssigned = job.assignedEmployees.employeeIds.some(
+      (id) => id.toString() === jobAssignDto.employeeToAssignId.toString(),
+    );
+    if (!alreadyAssigned) throw new ConflictException('Employee Not Assigned');
+    /// Role-based stuff
+    //TODO: Implement later
+
+    return await this.jobRepository.unassignEmployee(
+      jobAssignDto.employeeToAssignId,
+      jobAssignDto.jobId,
+    );
+  }
+
+  private async userIdMatchesEmployeeId(
     userId: Types.ObjectId,
     employeeId: Types.ObjectId,
-    jobId: Types.ObjectId,
   ) {
-    ///Validation
     const userExists = await this.usersService.userIdExists(userId);
     if (!userExists) throw new NotFoundException('User not found');
 
-    ///
-    //TODO: Implement later
-    console.log(employeeId);
-    console.log(jobId);
+    const employee: FlattenMaps<Employee> & { _id: Types.ObjectId } =
+      await this.employeeService.findById(employeeId);
+    if (!employee) throw new NotFoundException('Employee not found');
+    if (!employee.userId.equals(userId))
+      throw new UnauthorizedException('Inconsistent userId');
   }
 
   async assignEmployees(
@@ -430,24 +382,26 @@ export class JobService {
     jobAssignGroupDto: JobAssignGroupDto,
   ) {
     ///Validation
+    await this.userIdMatchesEmployeeId(userId, jobAssignGroupDto.employeeId);
 
-    //// User validation
     const job = await this.getJobById(jobAssignGroupDto.jobId);
-    for (const employeeId of jobAssignGroupDto.employeeIds) {
+    for (const employeeId of jobAssignGroupDto.employeesToAssignIds) {
       const exists = await this.employeeService.employeeExists(employeeId);
       if (!exists) {
         throw new NotFoundException('Employee not found');
       }
     }
     ///
-    const total = jobAssignGroupDto.employeeIds.length;
+    const total = jobAssignGroupDto.employeesToAssignIds.length;
 
     //remove duplicates
-    jobAssignGroupDto.employeeIds = [...new Set(jobAssignGroupDto.employeeIds)];
+    jobAssignGroupDto.employeesToAssignIds = [
+      ...new Set(jobAssignGroupDto.employeesToAssignIds),
+    ];
     let pass: number = 0;
 
     //const result = [];
-    for (const employeeId of jobAssignGroupDto.employeeIds) {
+    for (const employeeId of jobAssignGroupDto.employeesToAssignIds) {
       const isInJob = job.assignedEmployees.employeeIds.some((e) =>
         e.equals(employeeId),
       );
@@ -466,23 +420,15 @@ export class JobService {
     });
   }
 
-  async unassignEmployee(userId: Types.ObjectId, jobAssignDto: JobAssignDto) {
-    ///Validation
-    ///
-    return this.jobRepository.unassignEmployee(
-      jobAssignDto.employeeId,
-      jobAssignDto.jobId,
-    );
-  }
   async unassignEmployees(
     userId: Types.ObjectId,
     jobAssignGroupDto: JobAssignGroupDto,
   ) {
     ///Validation
+    await this.userIdMatchesEmployeeId(userId, jobAssignGroupDto.employeeId);
 
-    //// User validation //TODO
     const job = await this.getJobById(jobAssignGroupDto.jobId);
-    for (const employeeId of jobAssignGroupDto.employeeIds) {
+    for (const employeeId of jobAssignGroupDto.employeesToAssignIds) {
       const exists = await this.employeeService.employeeExists(employeeId);
       if (!exists) {
         throw new NotFoundException('Employee not found');
@@ -490,7 +436,7 @@ export class JobService {
     }
     ///
 
-    for (const employeeId of jobAssignGroupDto.employeeIds) {
+    for (const employeeId of jobAssignGroupDto.employeesToAssignIds) {
       const isInJob = job.assignedEmployees.employeeIds.some((e) =>
         e.equals(employeeId),
       );
@@ -507,8 +453,18 @@ export class JobService {
   }
 
   async getAllTagsInCompany(userId: Types.ObjectId, companyId: Types.ObjectId) {
-    const company = await this.companyService.companyIdExists(companyId);
-    if (!company) throw new NotFoundException('Company not found');
+    /// Validation
+    const user = await this.usersService.getUserById(userId);
+    if (!user) throw new NotFoundException('User not found');
+
+    const isInCompany = user.joinedCompanies.some(
+      (x) => x.companyId.toString() === companyId.toString(),
+    );
+    if (!isInCompany) throw new UnauthorizedException('User not in Company');
+
+    const companyExists = await this.companyService.companyIdExists(companyId);
+    if (!companyExists) throw new NotFoundException('Company not found');
+    ///
     return this.jobTagRepository.findAllTagsInCompany(companyId);
   }
 
@@ -516,6 +472,18 @@ export class JobService {
     userId: Types.ObjectId,
     companyId: Types.ObjectId,
   ) {
+    /// Validation
+    const user = await this.usersService.getUserById(userId);
+    if (!user) throw new NotFoundException('User not found');
+
+    const isInCompany = user.joinedCompanies.some(
+      (x) => x.companyId.toString() === companyId.toString(),
+    );
+    if (!isInCompany) throw new UnauthorizedException('User not in Company');
+
+    const companyExists = await this.companyService.companyIdExists(companyId);
+    if (!companyExists) throw new NotFoundException('Company not found');
+    ///
     return this.jobTagRepository.findAllPriorityTagsInCompany(companyId);
   }
 
@@ -523,12 +491,27 @@ export class JobService {
     userId: Types.ObjectId,
     createTagDto: CreateTagDto,
   ): Promise<boolean> {
+    /// Validation
+    const user = await this.usersService.getUserById(userId);
+    if (!user) throw new NotFoundException('User not found');
+
+    const isInCompany = user.joinedCompanies.some(
+      (x) => x.companyId.toString() === createTagDto.companyId.toString(),
+    );
+    if (!isInCompany) throw new UnauthorizedException('User not in Company');
+
+    const companyExists = await this.companyService.companyIdExists(
+      createTagDto.companyId,
+    );
+    if (!companyExists) throw new NotFoundException('Company not found');
+    ///
+
     const newTag = new JobTag(
       createTagDto.label,
       createTagDto.colour,
       createTagDto.companyId,
     );
-    const savedDoc = this.jobTagRepository.addJobTagToCompany(newTag);
+    const savedDoc = await this.jobTagRepository.addJobTagToCompany(newTag);
     console.log(savedDoc);
     return savedDoc != null;
   }
@@ -537,13 +520,30 @@ export class JobService {
     userId: Types.ObjectId,
     createPriorityTagDto: CreatePriorityTagDto,
   ): Promise<boolean> {
+    /// Validation
+    const user = await this.usersService.getUserById(userId);
+    if (!user) throw new NotFoundException('User not found');
+
+    const isInCompany = user.joinedCompanies.some(
+      (x) =>
+        x.companyId.toString() === createPriorityTagDto.companyId.toString(),
+    );
+    if (!isInCompany) throw new UnauthorizedException('User not in Company');
+
+    const companyExists = await this.companyService.companyIdExists(
+      createPriorityTagDto.companyId,
+    );
+    if (!companyExists) throw new NotFoundException('Company not found');
+    ///
+
     const newTag = new JobPriorityTag(
       createPriorityTagDto.label,
       createPriorityTagDto.priorityLevel,
       createPriorityTagDto.colour,
       createPriorityTagDto.companyId,
     );
-    const savedDoc = this.jobTagRepository.addJobPriorityTagToCompany(newTag);
+    const savedDoc =
+      await this.jobTagRepository.addJobPriorityTagToCompany(newTag);
     console.log(savedDoc);
     return savedDoc != null;
   }
@@ -552,6 +552,21 @@ export class JobService {
     userId: Types.ObjectId,
     deleteTagDto: DeleteTagDto,
   ) {
+    /// Validation
+    const user = await this.usersService.getUserById(userId);
+    if (!user) throw new NotFoundException('User not found');
+
+    const isInCompany = user.joinedCompanies.some(
+      (x) => x.companyId.toString() === deleteTagDto.companyId.toString(),
+    );
+    if (!isInCompany) throw new UnauthorizedException('User not in Company');
+
+    const companyExists = await this.companyService.companyIdExists(
+      deleteTagDto.companyId,
+    );
+    if (!companyExists) throw new NotFoundException('Company not found');
+    ///
+    //TODO: Cascade delete
     const deleteResult = await this.jobTagRepository.deleteJobTag(
       deleteTagDto.tagId,
     );
@@ -562,6 +577,21 @@ export class JobService {
     userId: Types.ObjectId,
     deleteTagDto: DeleteTagDto,
   ) {
+    /// Validation
+    const user = await this.usersService.getUserById(userId);
+    if (!user) throw new NotFoundException('User not found');
+
+    const isInCompany = user.joinedCompanies.some(
+      (x) => x.companyId.toString() === deleteTagDto.companyId.toString(),
+    );
+    if (!isInCompany) throw new UnauthorizedException('User not in Company');
+
+    const companyExists = await this.companyService.companyIdExists(
+      deleteTagDto.companyId,
+    );
+    if (!companyExists) throw new NotFoundException('Company not found');
+    ///
+    //TODO: Cascade delete
     const deleteResult = await this.jobTagRepository.deletePriorityTag(
       deleteTagDto.tagId,
     );
@@ -569,8 +599,7 @@ export class JobService {
   }
 
   async addCommentToJob(userId: Types.ObjectId, addCommentDto: AddCommentDto) {
-    const userExists = await this.usersService.userIdExists(userId);
-    if (!userExists) throw new NotFoundException('User not found');
+    await this.userIdMatchesEmployeeId(userId, addCommentDto.employeeId);
 
     const jobExists = await this.jobRepository.exists(addCommentDto.jobId);
     if (!jobExists) throw new NotFoundException('Job not found');
@@ -585,8 +614,7 @@ export class JobService {
     userId: Types.ObjectId,
     removeCommentDto: RemoveCommentDto,
   ) {
-    const userExists = await this.usersService.userIdExists(userId);
-    if (!userExists) throw new NotFoundException('User not found');
+    await this.userIdMatchesEmployeeId(userId, removeCommentDto.employeeId);
 
     const jobExists = await this.jobRepository.exists(removeCommentDto.jobId);
     if (!jobExists) throw new NotFoundException('Job not found');
@@ -601,8 +629,7 @@ export class JobService {
     userId: Types.ObjectId,
     updateCommentDto: UpdateCommentDto,
   ) {
-    const user = this.usersService.getUserById(userId);
-    if (!user) throw new NotFoundException('User not found');
+    await this.userIdMatchesEmployeeId(userId, updateCommentDto.employeeId);
 
     const job = await this.getJobById(updateCommentDto.jobId);
     if (!job) throw new InternalServerErrorException('JobId not found');
@@ -619,6 +646,6 @@ export class JobService {
       updateCommentDto.comment,
     );
     console.log(updateResult);
-    return true;
+    return updateResult;
   }
 }
